\name{helpRfuncs}
\alias{helpRfuncs}
\alias{aitoff_base}
\alias{analytic_nz}
\alias{analytic_nz_schec}
\alias{as.time}
\alias{colWeightedCounts}
\alias{colWeightedTabulates}
\alias{contour}
\alias{densityf}
\alias{dev2png}
\alias{draw.ellipses}
\alias{draw.sticks}
\alias{duplicated}
\alias{ecl2eq}
\alias{elementshift}
\alias{encompass}
\alias{eq2ecl}
\alias{eq2gal}
\alias{fft.convolve}
\alias{filename.append}
\alias{fround}
\alias{fsignif}
\alias{get_types}
\alias{hanning.smooth}
\alias{hist2D}
\alias{histf}
\alias{magaxis}
\alias{magbar}
\alias{maglim_weight}
\alias{make_grid}
\alias{open_plot}
\alias{plot.chain}
\alias{plot_covcor}
\alias{plot.cone}
\alias{plot.dens}
\alias{plot_hist2D}
\alias{plotCone}
\alias{read.chain}
\alias{read.colnames}
\alias{read.file}
\alias{running_stats}
\alias{scatter.plot}
\alias{showbw}
\alias{smooth.im}
\alias{smooth.im.par}
\alias{sph2sph}
\alias{split.expr}
\alias{tension_metrics}
\alias{text.coord}
\alias{triplot}
\alias{tround}
\alias{vecsplit}
\alias{vgsub}
\alias{weighted.quantile}
\alias{weighted.sd}
\alias{weighted.var}
\alias{write.file}
\alias{wtd_ecdf}
\title{Utility Functions for Scientific R Workflows}
\description{
  \pkg{helpRfuncs} collects reusable functions for scientific data analysis,
  visualization, file handling, and astronomical coordinate calculations.
}
\details{
  Plotting helpers include \code{magaxis}, \code{magbar}, \code{hist2D},
  \code{contour}, \code{triplot}, and coordinate-aware visualizations.
  Statistical tools include weighted summaries, density estimates, running
  statistics, and tension metrics. File helpers support tabular, FITS,
  Feather, and Parquet workflows. Coordinate utilities convert among
  equatorial, ecliptic, galactic, and spherical coordinate systems.
}
\seealso{
  See the package README for installation instructions, examples, and a
  feature overview.
}
\keyword{utilities}
